Some bug fixes from ARJ
Q-value panel was crashing, due to log10(q-value) giving infinity values. The plot is now just q-value rather than log10(q).

Also changed to falseDiscoveryRate code in mzidLib, new constructor allowing the code to rely on the isDecoy attribute rather than the manual decoy checkbox - previously this was non-functionally.

Added a progress bar for waiting for the graphs to load.